Decoder input dropout: replace with random codes (experimental)
Adding another flavor of decoder input dropout: instead of replacing the dropped-out input token with a mask token, replace it with a random token. The idea is to teach the decode not to fully trust past tokens even if they are not specifically a mask token, which is the case at inference.

To enable, set:
+model.data.dropout_decoder_random_id=true